pyproject.toml
[project]
name = "marketing"
version = "0.1.0"
description = ""
authors = [
    {name = "Gal Gilor", email = "ggilor@66degrees.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk (>=0.5.0,<0.6.0)",
    "google-genai (>=1.16.1,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "google-cloud-aiplatform[adk,agent-engines] (>=1.93.1,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]

[tool.poetry]
packages = [{include = "MarketingAgent"}]
package-mode = true


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
